Conflicts:
	cms/static/cms/js/plugins/cms.toolbar.js

+++ cms/admin/pageadmin.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from distutils.version import LooseVersion
 from functools import wraps
 import sys
 from cms.admin.placeholderadmin import PlaceholderAdmin
@@ -31,21 +30,22 @@ from cms.utils.compat.urls import unquote
 from cms.utils.helpers import find_placeholder_relation
 from cms.admin.change_list import CMSChangeList
 from cms.admin.dialog.views import get_copy_dialog
-from cms.admin.forms import PageForm, PageTitleForm, AdvancedSettingsForm, PagePermissionForm
+from cms.admin.forms import (PageForm, AdvancedSettingsForm, PagePermissionForm,
+                             PublicationDatesForm)
 from cms.admin.permissionadmin import (PERMISSION_ADMIN_INLINES, PagePermissionInlineAdmin, ViewRestrictionInlineAdmin)
 from cms.admin.views import revert_plugins
 from cms.models import Page, Title, CMSPlugin, PagePermission, PageModeratorState, EmptyTitle, GlobalPagePermission, \
-    titlemodels
+    titlemodels, StaticPlaceholder
 from cms.models.managers import PagePermissionsPermissionManager
-from cms.utils import helpers, moderator, permissions, get_language_from_request, admin as admin_utils, cms_static_url, copy_plugins
+from cms.utils import helpers, moderator, permissions, get_language_from_request, admin as admin_utils, copy_plugins
 from cms.utils.i18n import get_language_list, get_language_tuple, get_language_object
 from cms.utils.page_resolver import is_valid_url
 from cms.utils.admin import jsonify_request
 
 from cms.utils.permissions import has_global_page_permission, has_generic_permission
 from cms.utils.plugins import current_site
+from cms.utils.compat import DJANGO_1_4
 
-DJANGO_1_4 = LooseVersion(django.get_version()) < LooseVersion('1.5')
 require_POST = method_decorator(require_POST)
 
 if 'reversion' in settings.INSTALLED_APPS:
@@ -81,11 +81,11 @@ else:  # pragma: no cover
 
             return do_revision_context
 
-
     def create_revision():
         return ReversionContext()
 
 PUBLISH_COMMENT = "Publish"
+INITIAL_COMMENT = "Initial version."
 
 
 class PageAdmin(PlaceholderAdmin, ModelAdmin):
@@ -96,6 +96,7 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
     add_general_fields = ['title', 'slug', 'language', 'template']
     change_list_template = "admin/cms/page/tree/base.html"
     list_filter = ['published', 'in_navigation', 'template', 'changed_by', 'soft_root']
+    title_frontend_editable_fields = ['title', 'menu_title', 'page_title']
 
     inlines = PERMISSION_ADMIN_INLINES
 
@@ -109,8 +110,9 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
 
         url_patterns = patterns(
             '',
-            pat(r'^([0-9]+)/([a-z\-]+)/edit-title/$', self.edit_title),
+            pat(r'^([0-9]+)/([a-z\-]+)/edit-field/$', self.edit_title_fields),
             pat(r'^([0-9]+)/advanced-settings/$', self.advanced),
+            pat(r'^([0-9]+)/dates/$', self.dates),
             pat(r'^([0-9]+)/permission-settings/$', self.permissions),
             pat(r'^([0-9]+)/delete-translation/$', self.delete_translation),
             pat(r'^([0-9]+)/move-page/$', self.move_page),
@@ -142,6 +144,8 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
 
     def get_revision_instances(self, request, object):
         """Returns all the instances to be used in the object's revision."""
+        if isinstance(object, Title):
+            object = object.page
         placeholder_relation = find_placeholder_relation(object)
         data = [object]
         filters = {'placeholder__%s' % placeholder_relation: object}
@@ -183,7 +187,12 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
                 obj.rght = old_obj.rght
                 obj.lft = old_obj.lft
                 obj.tree_id = old_obj.tree_id
+        new = False
+        if not obj.pk:
+            new = True
         obj.save()
+        if new and Page.objects.filter(site_id=obj.site_id).count() == 1:
+            obj.publish()
         if 'recover' in request.path or 'history' in request.path:
             obj.pagemoderatorstate_set.all().delete()
             moderator.page_changed(obj, force_moderation_action=PageModeratorState.ACTION_CHANGED)
@@ -217,6 +226,8 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
             form = super(PageAdmin, self).get_form(request, obj, form=AdvancedSettingsForm, **kwargs)
         elif "permission" in request.path:
             form = super(PageAdmin, self).get_form(request, obj, form=PagePermissionForm, **kwargs)
+        elif "dates" in request.path:
+            form = super(PageAdmin, self).get_form(request, obj, form=PublicationDatesForm, **kwargs)
         else:
             form = super(PageAdmin, self).get_form(request, obj, form=PageForm, **kwargs)
         if 'language' in form.base_fields:
@@ -264,6 +275,9 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
         if not page.has_advanced_settings_permission(request):
             raise PermissionDenied("No permission for editing advanced settings")
         return self.change_view(request, object_id, extra_context={'advanced_settings': True, 'title': _("Advanced Settings")})
+
+    def dates(self, request, object_id):
+        return self.change_view(request, object_id, extra_context={'publishing_dates': True, 'title': _("Publishing dates")})
 
     def permissions(self, request, object_id):
         page = get_object_or_404(Page, pk=object_id)
@@ -385,7 +399,9 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
         context.update({
             'language': language,
             'language_tabs': languages,
-            'show_language_tabs': len(list(languages)) > 1,
+            # Dates are not language dependent, thus we hide the language
+            # selection bar: the language is forced through the form class
+            'show_language_tabs': len(list(languages)) > 1 and not context.get('publishing_dates', False),
         })
         return context
 
@@ -538,8 +554,8 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
                           }
                 helpers.make_revision_with_plugins(page, request.user, message)
 
-    def post_move_plugin(self, request, plugin):
-        page = plugin.placeholder.page
+    def post_move_plugin(self, request, source_placeholder, target_placeholder, plugin):
+        page = target_placeholder.page
         if page and 'reversion' in settings.INSTALLED_APPS:
             moderator.page_changed(page, force_moderation_action=PageModeratorState.ACTION_CHANGED)
             helpers.make_revision_with_plugins(page, request.user, _(u"Plugins were moved"))
@@ -910,20 +926,41 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
     @transaction.commit_on_success
     @create_revision()
     def publish_page(self, request, page_id):
-        page = get_object_or_404(Page, id=page_id)
+        try:
+            page = Page.objects.get(id=page_id, publisher_is_draft=True)
+        except Page.DoesNotExist:
+            page = None
         # ensure user has permissions to publish this page
-        if not page.has_publish_permission(request):
-            return HttpResponseForbidden(_("You do not have permission to publish this page"))
-        published = page.publish()
-        messages.info(request, _('The page "%s" was successfully published.') % page)
-        if "reversion" in settings.INSTALLED_APPS:
+        all_published = True
+        if page:
+            if not page.has_publish_permission(request):
+                return HttpResponseForbidden(_("You do not have permission to publish this page"))
+            published = page.publish()
+            if not published:
+                all_published = False
+        statics = request.GET.get('statics', '')
+        if not statics and not page:
+            return Http404("No page or stack found for publishing.")
+        if statics:
+            static_ids = statics .split(',')
+            for pk in static_ids:
+                static_placeholder = StaticPlaceholder.objects.get(pk=pk)
+                published = static_placeholder.publish(request)
+                if not published:
+                    all_published = False
+        if all_published:
+            messages.info(request, _('The content was successfully published.'))
+        else:
+            messages.warning(request, _("There was a problem publishing your content"))
+        if "reversion" in settings.INSTALLED_APPS and page:
             # delete revisions that are not publish revisions
             from reversion.models import Version
 
             content_type = ContentType.objects.get_for_model(Page)
-            versions_qs = Version.objects.filter(type=1, content_type=content_type, object_id_int=page.pk)
+            # reversion 1.8+ removes type field, revision filtering must be based on comments
+            versions_qs = Version.objects.filter(content_type=content_type, object_id_int=page.pk)
             deleted = []
-            for version in versions_qs.exclude(revision__comment__exact=PUBLISH_COMMENT):
+            for version in versions_qs.exclude(revision__comment__in=(INITIAL_COMMENT,  PUBLISH_COMMENT)):
                 if not version.revision_id in deleted:
                     revision = version.revision
                     revision.delete()
@@ -946,9 +983,12 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
         referrer = request.META.get('HTTP_REFERER', '')
         path = '../../'
         if 'admin' not in referrer:
-            if published:
-                public_page = Page.objects.get(publisher_public=page.pk)
-                path = '%s?edit_off' % public_page.get_absolute_url()
+            if all_published:
+                if page:
+                    public_page = Page.objects.get(publisher_public=page.pk)
+                    path = '%s?edit_off' % public_page.get_absolute_url()
+                else:
+                    path = '%s?edit_off' % referrer
             else:
                 path = '/?edit_off'
 
@@ -1156,14 +1196,30 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
             return True
         return super(PageAdmin, self).lookup_allowed(key, *args, **kwargs)
 
-    def edit_title(self, request, page_id, language):
+    def edit_title_fields(self, request, page_id, language):
         title = Title.objects.get(page_id=page_id, language=language)
         saved_successfully = False
+        raw_fields = request.GET.get("edit_fields", 'title')
+        edit_fields = [field for field in raw_fields.split(",") if field in self.title_frontend_editable_fields]
         cancel_clicked = request.POST.get("_cancel", False)
         opts = Title._meta
+
+        if not edit_fields:
+            # Defaults to title
+            edit_fields = ('title',)
+
         if not has_generic_permission(title.page.pk, request.user, "change",
                                       title.page.site.pk):
             return HttpResponseForbidden(_("You do not have permission to edit this page"))
+
+        class PageTitleForm(django.forms.ModelForm):
+            """
+            Dynamic form showing only the fields to be edited
+            """
+            class Meta:
+                model = Title
+                fields = edit_fields
+
         if not cancel_clicked and request.method == 'POST':
             form = PageTitleForm(instance=title, data=request.POST)
             if form.is_valid():
@@ -1173,7 +1229,7 @@ class PageAdmin(PlaceholderAdmin, ModelAdmin):
                 saved_successfully = True
         else:
             form = PageTitleForm(instance=title)
-        admin_form = AdminForm(form, fieldsets=[(None, {'fields': ('title',)})], prepopulated_fields={},
+        admin_form = AdminForm(form, fieldsets=[(None, {'fields': edit_fields})], prepopulated_fields={},
                                model_admin=self)
         media = self.media + admin_form.media
         context = {
